% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{bmi Body Mass Index of an individual}
\usage{
bmi(mass, height)
}
\arguments{
\item{mass}{in kg}

\item{height}{in m}
}
\value{
bmi as mass/(height * height)
}
\description{
bmi Body Mass Index of an individual
}
\examples{
bmi(mass = 80, height = 1.80)
bmi(mass = NA, height = 1.80)
bmi(mass = 80, height = NA)
bmi(mass = NA, height = NA)
}
