% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christinehoogland-package.R
\docType{package}
\name{christinehoogland-package}
\alias{christinehoogland}
\alias{christinehoogland-package}
\title{christinehoogland: Certification Level 2 - Developer - Exam}
\description{
French Certification of Professional Qualification - Level 2 - Developer – Package creation.
}
\author{
\strong{Maintainer}: Christine Hoogland \email{christine.hoogland@finddx.org} (\href{https://orcid.org/0000-0002-7341-1551}{ORCID})

Other contributors:
\itemize{
  \item FindDx [copyright holder]
}

}
\keyword{internal}
