% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_weight.R
\docType{data}
\name{data_weight}
\alias{data_weight}
\title{Population physical data}
\format{
A data frame with 87 rows and 3 variables:
\describe{
\item{name}{name of the character}
\item{mass}{mass in kg}
\item{height}{height in m}
...
}
}
\usage{
data_weight
}
\description{
A dataset containing mass and height for some characters.
}
\keyword{datasets}
